Add inverse lazy/weak check and change u64 address to u32 offset

+++ viking/src/checks.rs
@@ -175,7 +175,7 @@ pub struct FunctionChecker<'a, 'functions, 'orig_elf, 'decomp_elf> {
     decomp_addr_to_name_map: Lazy<elf::AddrToNameMap<'decomp_elf>>,
 
     known_data_symbols: KnownDataSymbolMap,
-    known_functions: FxHashMap<u64, &'functions functions::Info>,
+    known_functions: FxHashMap<u32, &'functions functions::Info>,
 
     pub orig_elf: &'orig_elf elf::OwnedElf,
     orig_got_section: Option<&'orig_elf goblin::elf::SectionHeader>,
@@ -417,7 +417,7 @@ impl<'a, 'functions, 'orig_elf, 'decomp_elf>
 
     /// Returns None on success and a MismatchCause on failure.
     fn check_function_call(&self, orig_addr: u64, decomp_addr: u64) -> Option<MismatchCause> {
-        let info = *self.known_functions.get(&orig_addr)?;
+        let info = *self.known_functions.get(&(orig_addr as u32))?;
         let name = info.name().as_str();
         let decomp_symbol = self.decomp_symtab.get(name)?;
         let expected = decomp_symbol.st_value;
